pyproject.toml
[project]
name = "recsys"
version = "0.0.1"
description = "A production-quality hybrid recommendation system combining **RAG (Retrieval-Augmented Generation)** with **Collaborative Filtering** to recommend board games. Built as a portfolio project showcasing end-to-end ML engineering skills."
authors = [{ name = "Hanish Paturi", email = "hanishpaturi1320@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.11,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "alembic>=1.17.1",
    "dask>=2025.11.0",
    "dbt-core>=1.10.15",
    "dbt-snowflake>=1.10.3",
    "dvc>=3.63.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "implicit>=0.7.2",
    "ipykernel>=7.1.0",
    "ipynb>=0.5.1",
    "jq>=1.10.0",
    "kaggle>=1.7.4.5",
    "loguru>=0.7.3",
    "mlflow>=3.6.0",
    "numpy",
    "pandas",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "sentence-transformers>=5.1.2",
    "snowflake>=1.8.0",
    "sqlalchemy>=2.0.44",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
]

[project.urls]
Homepage = "https://hantablack9.github.io/gamerecommendationapp/"
Repository = "https://github.com/hantablack9/gamerecommendationapp"
Documentation = "https://hantablack9.github.io/gamerecommendationapp/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.23.0",
    "ty>=0.0.1a16",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "mypy>=1.18.2",
    "pytest-asyncio>=1.3.0",
    "black>=25.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/recsys"]

# [tool.hatch.build.sources]
# "src/components" = "multivariate_temp_forecast"

# [tool.hatch.build.sources]
# "src" = "."
# --- END OF ADDED SECTION ---

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true


[tool.ty.environment]
python = "./.venv"
python-version = "3.12"


[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true
src = ["src"]


[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    "TRY003",
    "TRY300"
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]
